requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.13
scikit-learn>=1.5
xgboost>=2.0
pybaseball>=2.2
MLB-StatsAPI>=1.7
requests>=2.32
python-dotenv>=1.0
pyarrow>=16.0
joblib>=1.4
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.9
streamlit>=1.35
ipykernel>=6.29
jupyterlab>=4.2
pytz>=2024.1
